+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2 - in progress
 - queue added
+- duckduckgo search added
 
 ### 0.0.1 - 2021.07.10